+++ app/components/SideBar.js
@@ -1,40 +1,32 @@
 import React from 'react';
 import TreeView from './treeView';
+var classNames = require('classnames');
 import $ from 'jquery';
 
 
 var SideBar = React.createClass ({
-	getInitialState: function() {
-    	return {results: []};
+	getInitialState() {
+    	return {isExpanded: false, results: []};
   	},
-	componentDidMount: function()
-	{
-		$.ajax({
-			type: "GET",
-			dataType: "json",
-			url: "http://tadadev.klp.org.in/api/v1/boundaries/",//TODO: Make a call that fetches only schools and districts
-			success: function(data) {
-						console.log(data.results);
-						this.setState( {
-							results: data.results
-						});
-					}.bind(this) //end of success callback
-		});//end of ajax block
-	},
+  toggleTree() {
+    this.setState({isExpanded: !this.state.isExpanded})
+  },
+	componentDidMount() {
+
+  },
 
 	render: function() {
+  var sidebarClass = classNames({
+  'toggled': this.state.isExpanded})
 		return (
-			<div id="wrapper" className="">
          <div id="sidebar-wrapper">
             <div id="treetoggler">
               <a href="#menu-toggle" className="btn btn-primary btn-xs" id="menu-toggle">
-                <span id="toggler-icon" className="glyphicon glyphicon-resize-horizontal"></span>
+                <span id="toggler-icon" onClick={this.toggleTree} className="glyphicon glyphicon-resize-horizontal"></span>
               </a>
             </div>
           <div id="treeview_side" className="treeview"><TreeView/></div>
         </div>
-      </div>
-
 		);
 	}
 });
