+++ app/components/treeView.js
@@ -26,15 +26,32 @@ const dataSource = [
 // [controlled component](http://facebook.github.io/react/docs/forms.html#controlled-components)
 // is preferred.
 const CompanyPeople = React.createClass({
+  getInitialState() {
+      return {results: []};
+    },
+
+  componentDidMount() {
+    $.ajax({
+      type: "GET",
+      dataType: "json",
+      url: "http://tadadev.klp.org.in/api/v1/boundaries/",//TODO: Make a call that fetches only schools and districts
+      success: function(data) {
+            console.log(data.results);
+            this.setState( {
+              results: data.results
+            });
+          }.bind(this)
+    });
+  },
   render() {
     return (
       <div>
-        {dataSource.map((node, i) => {
-          const type = node.type;
-          const label = <span className="node">{type}</span>;
+        {this.state.results.map((boundary, i) => {
+          const name = boundary.name;
+          const label = <span className="node">{name}</span>;
           return (
-            <TreeView key={type + '|' + i} nodeLabel={label} defaultCollapsed={false}>
-              {node.people.map(person => {
+            <TreeView key={name + '|' + i} nodeLabel={label} defaultCollapsed={false}>
+              {/*node.people.map(person => {
                 const label2 = <span className="node">{person.name}</span>;
                 return (
                   <TreeView nodeLabel={label2} key={person.name} defaultCollapsed={false}>
@@ -43,7 +60,7 @@ const CompanyPeople = React.createClass({
                     <div className="info">role: {person.role}</div>
                   </TreeView>
                 );
-              })}
+              })*/}
             </TreeView>
           );
         })}
