+++ package.json
@@ -5,6 +5,7 @@
   "main": "webpack.config.js",
   "start": "node server.js",
   "dependencies": {
+    "classnames": "^2.2.0",
     "flux": "^2.1.1",
     "jquery": "^2.1.4",
     "keymirror": "^0.1.1",
